requirements.txt
asgiref==3.3.1
attrs==20.3.0
coverage==5.3
Django==3.1.4
django-countries==7.0
django-crispy-forms==1.10.0
django-datatables-view==1.19.1
django-filter==2.4.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
factory-boy==3.1.0
Faker==5.0.1
iniconfig==1.1.1
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
